build: move storybook/mocks to test/fixtures
